Created script to run all population scrips
